Tests/LibWeb/Text/expected/canvas/pattern-from-image.txt

null
null
PASS (didn't throw!)
